Extract bootstrap and router from cmd/uniproxy/main.go (#2) (#33)
Pure restructuring with zero behavior change. main.go drops from 1,688 to
119 LOC by moving service wiring into internal/bootstrap and route
registration into internal/router.

- internal/bootstrap/bootstrap.go: Bootstrap(ctx, cfg) returning AppServices
  with every wired store, service, handler, kube client, sidecar manager,
  session manager.
- internal/bootstrap/{otel,registry_loader}.go: OTel init and registry
  loader helpers. The loader duplicates internal/handlers/registry.go's
  copy on purpose; issue #3 consolidates both into pkg/services/registry.
- internal/router/router.go: Register(r, svc) installing CORS, /health,
  /docs, and the full /api/v1 surface.
- internal/router/mcp_rest.go: REST-style MCP handlers (tools/resources/
  prompts list/call/read/get) plus their session and auth helpers, moved
  alongside their only callers.
- Dropped four dead helpers from main.go (generateID local duplicate,
  isVirtualMCPAdapter, reconfigureVirtualMCPAdapter, handleMCPProxy) with
  no remaining callers.

Verified: go build ./... clean, go vet clean on new packages, binary
boots, /health /auth/mode /api/v1/registry /api/v1/adapters
/docs/index.html all 200, graceful shutdown intact. Pre-existing test
failures in pkg/mcp, internal/service, and two flaky tests confirmed
present on main.
